1、增加支持 FIL 相关接口; * cpp增加依赖库 submodules uint256_t;

+++ app/src/main/cpp/include/utility/util.h
@@ -63,7 +63,7 @@ std::string Join(std::vector<std::string> v, std::string pattern);
  */
 std::string DecStringToHexString(std::string str);
 
-std::string numberToHexString(unsigned int val);
+std::string numberToHexString(unsigned long val);
 
 /*
  * char array<uint8_t, 32> -> Hexadecimal string conversion
